+++ pom.xml
@@ -9,7 +9,7 @@
 	</parent>
 	<groupId>jp.igapyon</groupId>
 	<artifactId>simpleodata4</artifactId>
-	<version>0.20210316.1</version>
+	<version>0.20210316.2</version>
 	<name>simpleodata4</name>
 	<description>Simple Odata4 sample.</description>
 	<properties>
