+++ src/main/java/jp/igapyon/simpleodata4/oiyokan/OiyokanConstants.java
@@ -1,0 +1,8 @@
+package jp.igapyon.simpleodata4.oiyokan;
+
+public class OiyokanConstants {
+    /**
+     * Oiyokanのバージョン番号
+     */
+    public static final String VERSION = "20210316b";
+}
