+++ src/main/java/jp/igapyon/simpleodata4/oiyokan/OiyokanOdata4Register.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License. 
  */
-package jp.igapyon.simpleodata4.entity;
+package jp.igapyon.simpleodata4.oiyokan;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
