+++ src/main/java/jp/igapyon/simpleodata4/oiyokan/h2/data/TinyH2Util.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License. 
  */
-package jp.igapyon.simpleodata4.h2data;
+package jp.igapyon.simpleodata4.oiyokan.h2.data;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
