+++ src/main/java/jp/igapyon/simpleodata4/oiyokan/h2/sql/TinySqlBuildInfo.java
@@ -1,4 +1,4 @@
-package jp.igapyon.simpleodata4.h2data.sqlbuild;
+package jp.igapyon.simpleodata4.oiyokan.h2.sql;
 
 import java.util.ArrayList;
 import java.util.List;
