+++ src/main/java/jp/igapyon/simpleodata4/oiyokan/package-info.java
@@ -1,4 +1,4 @@
 /**
  * 要素のデータを構築するクラス群.
  */
-package jp.igapyon.simpleodata4.entity;
+package jp.igapyon.simpleodata4.oiyokan;
