+++ src/test/java/jp/igapyon/simpleodata4/SimpleOdata4AppTests.java
@@ -1,13 +1,10 @@
 package jp.igapyon.simpleodata4;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+// 【現状 Spring Boot のテストはない. ビルド時間短縮のためコメントアウト】 @SpringBootTest
 class SimpleOdata4AppTests {
-
     @Test
     void contextLoads() {
     }
-
 }
