+++ src/test/java/jp/igapyon/simpleodata4/oiyokan/h2/data/H2DatabaseTest.java
@@ -1,12 +1,15 @@
-package jp.igapyon.simpleodata4.h2data;
+package jp.igapyon.simpleodata4.oiyokan.h2.data;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.sql.Connection;
 
 import org.junit.jupiter.api.Test;
 
-// @SpringBootTest
+/**
+ * そもそも内部 h2 database への接続性を確認
+ */
 class H2DatabaseTest {
-
     @Test
     void test01() throws Exception {
         Connection conn = TinyH2Util.getH2Connection();
@@ -17,13 +20,10 @@ class H2DatabaseTest {
         // テーブルデータをセットアップ.
         TinyH2DbSample.setupTableData(conn);
 
-        try (var stmt = conn.prepareStatement("SELECT ID, Name, Description FROM MyProducts ORDER BY ID")) {
+        try (var stmt = conn.prepareStatement("SELECT ID, Name, Description FROM MyProducts ORDER BY ID LIMIT 3")) {
             stmt.executeQuery();
             var rset = stmt.getResultSet();
-            for (; rset.next();) {
-                // System.err.println(rset.getString(1) + ":" + rset.getString(2) + ":" +
-                // rset.getString(3));
-            }
+            assertEquals(true, rset.next());
         }
 
         conn.close();
